Run on cross-platform

+++ encoding.cpp
@@ -145,7 +145,7 @@ char * unicode_to_utf8(unsigned short unicode)
 int gbk_or_big5_2_utf8(char* in, const char* encoding)
 {
     CodeConverter cc = CodeConverter(encoding, "utf-8");
-    size_t outlen = strlen(in) * 2;
+    size_t outlen = static_cast<size_t>(strlen(in) * 2);
     char outbuf[outlen];
 
     ofstream out("convert-output.txt", ios_base::out | ios_base::ate);
